requirements.txt
python-dotenv==0.6.2
click==6.7
requests==2.13.0
retry==0.9.2

simplejson==3.10.0
